requirements.txt
pandas==0.24.2
tqdm==4.36.1
opencv-python==4.2.0.34
matplotlib==3.0.3
numpy==1.16.4
tensorflow-gpu==1.8.0
Pillow==6.1.0
pypng==0.0.20
pfm==0.6.0
scipy==1.1.0